Fix ftp-uri spec: use URI instead of URL object
The previous spec was failing when a ftps uri was provided.
The problem was that ftps:// is NOT a recognized protocol by
the .NET framework, so creating an URL object with it throwed
an exception. Instead, if we use URI, it doesnt't check if the
protocol is supported or not, so the spec validates correctly.