Return empty list of objects if path doesn't exist
When listing objects we were not checking if the cd operation
was succesfull, so we were returning the list of objects in
the root directiory if the specified path didn't exist.

The method returns an empty list of objects in order to
to mantain the compatibility with the S3 object-storage
implementation.
[Re #1] [Fix #1]